tbtadm: add `add` command
This allows the user to add a device to ACL without using `approve` or
`approve-all` commands, and thus allowing such addition even when the
device is already authorized (e.g. on pre-boot environment).
This can't work on SL2 because this is how the sysfs interface to the
driver is defined.

Resolves #34

Signed-off-by: Yehezkel Bernat <yehezkel.bernat@intel.com>